drastically improve performance using requestAnimationFrame and dirtyChecking
instead of calling on each event the callback, it calls max on each frame the callback.
Also set the width of sensors to extreme high value instead of read expensive dimensions from parent.

Compared to the <object> sensor this variant is faster in each frame and way faster in the initialization.

Fixes #85