Fix ElementQueries always attaching to window and listening
When we originally added UMD support, we forgot to move the logic that forcibly calls .listen() out of the block at the bottom and into the UMD logic block. This, in turn, meant that EQ was always being attached to window and listen was always being called, even when using EQ or ResizeSensor in a front-end environment with module support.