Adding error check for InvalidAccessError (#197)
This only happens with Firefox due to their security policy. 
Fix found in previous issue thread. https://github.com/marcj/css-element-queries/issues/117
Can confirm this was due to typekit being included in the project. 
It was quite annoying to debug in a build environment via es6 module loader (systemJS) so hopefully, this PR solves the issue for people. 
All credit to @jyrkij