Use ResizeObserver when available and check for initial invisible elements (#213)
ResizeObserver is Chrome only at the moment.

When a resize  sensor has been added to an invisible element, then we check coming with this commit every frame whether it is actually visible (rendered) and then start the sensor process. Otherwise the sensor would just not work on initial invisible elements.